backend/auth/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
pymongo==3.12.3
python-decouple==3.8
